Rename ArpaAnnualPerformanceReporter with View postfix
